server: Fix path decoding on SoupServer
SoupServer decodes the paths when raw-paths property is FALSE, but
calling soup_message_set_uri() with the decoded uri, encodes it again
because of the normalization. Since we know the uri is already
normalized, we just want to change the path, so I've added a private
function to set the uri of a SoupMessage without doing the
normalization.

Fixes server-auth-test